+++ lib/src/signatures.dart
@@ -1,6 +1,6 @@
 library just_jwt.signatures;
 
-/// Verifiers the [signature] of the [message].
+/// Verifies the [signature] of the [message].
 typedef bool Verifier(String message, List<int> signature);
 
 /// Returns a signature of the [message].