+++ CHANGELOG.md
@@ -11,3 +11,8 @@
 # 1.1.1
 
 - Works with non-padded BASE64 strings.
+
+# 1.2.0
+
+- Improved Signers/Verifiers composition.
+- New ```createJwaRS256Verifier``` function which creates a RSA256 verifier from encoded modulus and encoded exponent as described in RFC 7518.